+++ src/index.11tydata.js
@@ -1,0 +1,1 @@
+export { eleventyComputed } from "./updates/updates.11tydata.js";